add ingress for apache httpd service
The existence of the apache httpd service is a bit extraneous, because
Kubernetes already has an ingress controller which has an identical
role.

However, to keep local dev environments as similar as possible to the
production environment we will route the requests from the ingress
controller through the apache httpd service and then to the node
services.